refactor(export): extract shared module to break inline-assets import cycle
